mdb_test_auto/pyproject.toml
[tool.poetry]
name = "mdb-test-auto"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mdb_test_auto", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pyserial = "^3.5"
uvicorn = "^0.30.1"

[tool.poetry.scripts]
server = "server:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
